Fix regressions in local development (#45)
This PR fixes some regressions in #36 that broke local development. The main change is in how we construct URLs for the flink master — instead of passing around a string like "servicename.namespace" we construct the URL in kubernetes code directly from the application. This allows it to work for both proxy URLs (needed in local development) and cluster URLs and avoids the risk of string manipulation deeper in the program.