Simplify Flink config configuration management (#57)
This PR is an attempt to simplify how we manage Flink configuration (i.e., the stuff that ends up in flink-conf.yaml) in the operator. It also lays the groundwork for implementing STRMCMP-352.  There is also a companion PR https://github.com/lyft/streamingplatform/pull/589 to update the s2ibuilder.

Currently Flink configs are injected into the config file in 3 ways:
* As individual environmental substitutions in the [docker-entrypoint.sh](https://github.com/lyft/streamingplatform/blob/894be33f7617ffa2367ce3eea4563fc3be54f478/s2i-streamingplatform-flink/docker-entrypoint.sh)
* As a mass env substitution via FLINK_CONFIG_OVERRIDES
* By direct appending of the config file in docker-entrypoint.sh

and the config values themselves can come from:
* defaults in the flink-conf.yaml in the image
* the operator config (for heap size and savepoint/checkpoint dirs)
* individual configs in the CRD (ports mostly)
* the flinkConfigs field in the CRD

With this PR, configs are managed in just two places: there are defaults in the flink-conf of the image, and overrides are injected via a single environment variable OPERATOR_FLINK_CONFIG, whose value is appended to flink-conf.yaml in docker-entrypoint.

Configs in the operator come from two places; configs that are needed by the operator for it to function, or which come from kubernetes itself (taskmanager slots, various ports, the jobmanager rpc address) are defined as static fields in the CRD. All other configs are supplied optionally by the user in the flinkConfig map.

I believe that this will be easier for users to understand and harder to screw up. Currently the user can fairly easily break things by making apparently innocuous changes to the flink-conf.yaml in their image (e.g., by removing or modifying the jobmanager.rpc.address config or changing some of the port configs). With this PR, they can modify their flink-conf.yaml in any way (such that it remains a valid flink config file) and the operator will continue to work properly.

This also removes some of the "magic" currently in the operator or the s2i image, like the automatic configuration of checkpoint / savepoint dirs or of HA, and the special casing of seemingly arbitrary flink configs. Users will now be on the hook to configure those things themselves, which is currently an inconvenience, however that will be addressed in the near future by the streamingplatform jsonnet library.

@gfee-lyft: note that once this is deployed, you will need to manually configure the savepoint/checkpoint dirs, your heap settings, and HA config if have that set up:

```yaml
spec:
  flinkConfig:
    state.backend.fs.checkpointdir: s3:///lyft-modelbuilder/metadata/flinkk8operator/flink/{app}/checkpoints
    state.checkpoints.dir: s3:///lyft-modelbuilder/metadata/flinkk8operator/flink/{app}/externalized-checkpoints
    state.savepoints.dir: s3:///lyft-modelbuilder/metadata/flinkk8operator/flink/{app}/savepoints
```