[STRMCMP-365] Operator integration testing (#76)
This PR follows on #70 to add more sophisticated integration tests for the operator. See [integ/README.md](https://github.com/lyft/flinkk8soperator/blob/micah_integ2/integ/README.md) for more on these tests.

In order to simplify interaction with the custom resources in the tests, I've also added kubernetes client generation fro the CRD, and a bunch of helper functions in the integ/framework package.

In addition to a lot of general integration testing infrastructure, this PR adds tests for the following scenarios:
* Submitting a job, validating it successfully starts, updating it, then deleting it
* Submitting a job, causing it to starting failing, then attempting to update it
* Submitting a job which will timeout when it is snapshotting, and attempting to update it

Starting up the environment and running the tests takes ~9 minutes. Unfortunately, due to limitations in Travis (https://github.com/travis-ci/travis-ci/issues/5035) these will appear in the same github status as the build and unit tests. They must also currently run after the build has finished, as they depend on the image created as part of the build process. If this becomes a burden, we may decide to run these tests only on master builds. For now I think the slower build times are worth the large increase in confidence they give us in our changes.