[STRMCMP-352] Update flink applications on all config changes (#81)
Currently the operator will only respond to two changes of the CRD: a change to the image or a change to the parallelism, and only the former case will it actually recreate the cluster. That assumption is baked fairly deep into the code, for example deployments are named according the image.

This PR expands the set of changes that will trigger a deployment to almost all fields of the CRD (the sole exception is DeploymentMode, which has no effect on the cluster). This required a bunch of changes to how the operator functions, but also allowed for some nice simplification and code deletion.

The main changes:
* Deployments/pods/etc are now named with a hash derived from the deployment template, instead of just by the image
* We no longer rescale the deployments when parallelism is changed; instead we just do a normal deployment with the new parallelism. This reflects the fact that in flink, a rescale is really just a redeploy with a new parallelism.
* To determine if an application has changed, we compare the actual deployments to the ones we generate from the application. Note that this will mean that we need to be careful in the future with changes to how a flinkapplication is translated into a deployment. Changing that will cause re-deploys for all services, and should be considered a breaking change.