[STRMCMP-446] Use unique high-availability.cluster-id (#94)
Currently the high-availability.cluster-id is set by the user, typically to the job name. This can create conflicts if the job and cluster are not cleaned up properly or during a Dual mode deployment. While #93 helps ensure that jobs are cleaned up, we still want to use unique cluster ids to avoid any potential issues.

This PR injects a config for high-availability.cluster-id as `{app name}-{hash}`. This is awkward, because the hash itself depends on the flink config. So we must inject it after the hash is computed instead of including it in the initial config construction.