[STRMCMP-402] Cancel jobs before destroying cluster (#93)
This PR adds support for cancelling flink jobs before the cluster is torn down by kubernetes.

This is accomplished by adding a [finalizer](https://kubernetes.io/docs/tasks/access-kubernetes-api/custom-resources/custom-resource-definitions/#finalizers) to the custom resource, which prevents deletion until the finalizer is removed.

By default, we will now cancel-with-snapshot before deleting the resource and tearing down the cluster. However, in some situation this may not be possible (the job may be failing, snapshots may be timing out, or the flink cluster may be so wedged that no API requests are succeeding.) We need a way for the user to indicate the equivalent of `sp cancel -f`. Ideally we would be able to take advantage of Kubernete's built-in `deletionGracePeriodSeconds` to tell the operator how long it has to attempt a graceful delete, but until https://github.com/kubernetes/kubernetes/issues/56567 is fixed this is not supported for custom resources.

So instead we introduce a new field to the FlinkApplication spec: DeleteMode, which can take the following values:

* `Savepoint` (default) The operator will take a final savepoint before cancelling the job, and will not tear down the cluster until a savepoint has succeeded.
* `ForceCancel` The operator will force cancel the job before tearing down the cluster
* `None` The operator will immediately tear down the cluster

So if the savepoint is failing and the user is comfortable losing the state, they can change this value to either ForceCancel or None to allow the delete to go through. Our tooling (sp2) will support setting this automatically if `-f` is passed to the `delete` command.