Add Example application inside operator for test and reference (#96)
This PR - add a sample application that produces a docker image used in the integration tests.

The sample application also acts as an example for reference purposes, when individuals try to adopt Flink operator and try to build images to run their applications.

Note: Currently the docker image for `application_example` has been built manually and pushed to dockerhub. This image can be exposed to public. It has no lyft bits/business logic in it.