[STRMCMP-364] Implement new operator state machine (#101)
This PR implements the new operator state machine, as described [here](https://docs.google.com/document/d/1OTOvaC16m2zdPxWkj5qIC2rUZheKdrm1tlt9ji48k74/edit).

The primary goal of this redesign is to improve the uptime of jobs by allowing them to automatically recover and roll back from bad deploys, and to ensure that jobs never become unrecoverable by entering a "failed" state.

As part of this work (covered by STRMCMP-364), I've also addressed the following issues
* STRMCMP-342 (correctly handling failed savepoint restores)
* STRMCMP-431 (automatically rolling back failed deploys)
* STRMCMP-467 (both old and new jobmanagers were routed to by the service during dual deploys)

This is a very large PR. A few highlights to help guide reviewers:

## FlinkApplication resource changes
In order to support the new functionality, several fields have been added to the Status block of the FlinkApplication resource:
* `JarName`, `Parallelism`, `EntryClass`, and `ProgramArgs`: these track the job parameters of the currently running job, which are needed to re-submit and old job during rollback. They will also allow us to fix STRMCMP-473.
* `DeployHash`: this is the hash of the current FlinkApplication. It allows us to determine what cluster to rollback to during a deployment.
* `FailedDeployHash`: if set, this indicates to the operator that we have already attempted, and failed, to deploy the FlinkApplication with this hash. Until the spec changes to not match this hash, we will not re-attempt the deployment.

In addition, the `StoppedAt` field was removed, as it's no longer possible for FlinkApplications to stop.

## flink_state_machine.go
This has been largely rewritten to implement the new state machine. A guide to the new behavior can be found in docs/state_machine.md. 

## jobmanager services
In the current world we have just one service for the jobmanager (`{name}-jm`) which includes all jobmanager pods for the application. However, as described in STRMCMP-467, this can cause issues during Dual deployments, because we may attempt to cancel the job in the new jobmanager, or submit the job to the old one. Even worse, the taskmanagers may try to register with the wrong one. In order to ensure that we are always communicating with the right jobmanager, we now have two services: the generic one (renamed to `{name}` as the `-jm` suffix is unnecessary) and a version-specific one (`{name}-{hash}`). The versioned service only ever includes the pods for that hash, while the generic one will be switched to the new hash during the deploy process.

The taskmanagers are also now configured to communicate with the jobmanager through the versioned service.

As part of this work, a `hash` parameter has been added to all functions that communicate with the jobmanager so that we can target a particular service.

## Single mode
This PR removes support for the Single DeployMode. This was done to simplify this already very complex PR. We expect all of our users will want to use Dual mode given its additional safety and lower downtime. I've added TODOs in the code where different behavior for single mode needs to be added, and after this PR is merged I'll create a JIRA to track its re-implementation.