Migrate from Operator SDK to Controller Runtime (#102)
This PR removes the dependency from https://github.com/operator-framework/operator-sdk and now depends on https://github.com/kubernetes-sigs/controller-runtime. The customization for various use cases are all in Controller Runtime. The Operator SDK on the top currently does not provide added value. 

In order to get features from Controller runtime, we would be tied to Operator SDK to be upgraded. This limitation is no more.

All calls to get pods, deployments and services are now cached. We will add concurrent access wherever required. Currently the resources in the cache are scoped by namespace and Flink application owner references. 

This change also has support to configure multiple workers.

To be added:
- Support for multiple namespaces.