[STRMCMP-825] Improve error messaging when moving to DeployFailed in a non-recoverable state (#171)
Currently, all transitions to deployFailed production a Successfully rolled back deploy event (via the FlinkStateMachine.deployFailed method). However, there are some cases where the rollback in fact has failed; for example in the sequence where

1. cancelling the existing job succeeds, but we fail to find the savepoint path for some reason
2. there's never been a successful checkpoint taken for the job (or we fail to query it), so recovery fails

This change improves the error message in those cases to be clearer to the user that rollback has failed, and manual intervention is needed to recover.

Also includes a nil-pointer deference bug fix for the output of GetJobForApplication, which I encountered while reproducing this.