Use Go Modules (#274)
And while we're at it, upgrade to Go 1.20. In the process I had to
change a lot of the dependencies. Some because the dependencies wouldn't
resolve using Go modules, some because they were incompatible with
others. The long and the short of it is: 1) the kubernetes dependency
ecosystem is very rough and 2) this should be carefully validated during
rollout given the number of dependency shifts.

Some other minor notes include some lint changes, or rewriting a very
small bit of code to drop a dependency to have one less to fight with.

---------

Co-authored-by: Seth Saperstein <sethsaperstein@lyft.com>