Monitor Job Vertices State on Deploy (#284)
# Context
Currently flinkk8soperator only monitor Flink job level overview status,
and if the job status is RUNNING, it considers the deployment is
succeeded.

However we encountered scenarios that some vertices are in bad states,
the job keeps crash and restart. So the change is also check all
vertices status.

[Design doc](
https://docs.google.com/document/d/1m9dUKtIvR8NDZZ9le-CYGEmeuod20vXSL73OiQiP1j8/edit)

# Implementation 
1. ~~Use a feature flag to control the feature ON/OFF. Initially default
value is false. After we have enough confidence, we will flip the flag
or remove the feature flag.~~ As discussed with Anand, TLDR is this
change is relatively small and I will do handful manual testing in
staging after merge.
2. I add a logic to fail fast if there is any vertex status is `FAILED`.
3. Add several unit tests
4. I was trying to add a integration test which will add a `,` in
`programArgs` and assume it will break args parsing. However, currently
test flink app is pinned to a specific old image. I tried to build a new
flink image, but new image exceeded memory limit.

# Jira link
https://jira.lyft.net/browse/STRMCMP-1640